requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
phonenumbers==8.13.29
pymongo==4.6.1
numpy==1.26.3
requests==2.31.0